renaming localdocker.py to local.py
