+++ buildstockbatch/local.py
