+++ setup.py
@@ -89,7 +89,7 @@ setuptools.setup(
     },
     entry_points={
         'console_scripts': [
-            'buildstock_docker=buildstockbatch.localdocker:main',
+            'buildstock_local=buildstockbatch.local:main',
             'buildstock_eagle=buildstockbatch.eagle:user_cli',
             'buildstock_aws=buildstockbatch.aws.aws:main'
         ]
